+++ library/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
+- Add `nanoid` action to validate Nano IDs (pull request #789)
 - Add `undefinedable` and `undefinedableAsync` schema (issue #385)
 
 ## v0.39.0 (August 24, 2024)